audit: deserialize the questions JSON blob into proper models
This was a bit fiddly but it makes the interface a lot nicer. JSON is
parsed from the DB and cast into the corresponding
Audit::Question/Criterion classes which provide an easy way to work
with that unstructured data.

Note that the YAML input for the standards is now expected to store
the category alongside the question. Previously:

accessibilité:
  - title: bla
    id: bla
    criteria:
    - bla
vie-privée:
  # etc

Now a simpler array is expected in this shape:

- title: bla
  category: accessibilité
  id: bla
  criteria:
    - bla
- title: foo
  category: vie-privée
  # etc

This means one less level of nesting and the grouping can be done by
the libs instead of us.