audit: allow moving straight onto the next unanswered question
This is dodgy because we currently distinguish:

- answered: you've checked the criteria and clicked save;
- answered but incomplete: you didn't check the criteria but you
  clicked save and so we stored "0" as the answer, so it's actually
  answered (with no's)
- unanswered: you didn't save your answer, or went back.

and we always redirect to unanswered/unseen before.

This will be confusing to users because they will rightfully assume
"next question" can equally be what is not unanswered, or what is
still incomplete.

When we move to radio answers (explicit Yes/No/Not application choice,
or no choice at all), this will go away.