question: don't offer next question yet
If we're going to edit some quality standards, as we've done with
a11y (thanks to Anne-So), it's not really a goal to move fast to the
next question: taking the time to read and understand the
question/criteria is the focus, so adding an option to dash to the
next thing isn't worth it at the moment, and it adds visual noise too.

Moving to the "next" question is also a bit vague since the questions
aren't really ordered (alphanumericaly, accidentally) so you don't
really know where you'll end up.

Hide it behind an ENV variable which will be false by default until we
gather more requirements.