config: undo the staging environment
I'd written the new staging environment to differentiate the
credentials which are encrypted per env (i.e Rails decrypts based on
RAILS_ENV) but that also forces me to duplicate every other resource
that is read according to RAILS_ENV, like `database.yml`, `cache.yml`,
etc.

That's annoying and error-prone so instead revert to a single
environment and a single credentials file with a shadow
BETA_STANDARDS_ENV variable to pick the right credentials within.

This also makes the staging environment closer to production as the
RAILS_ENV differentiation disappears in the code.