proconnect: add a post logout action
We can't clear the session before sending them off ot
logout (clear_proconnect_setup!) because the Omniauth code reuses the
session to figure out the `id_token_hint` which needs to be passed in
the ProConnect logout endpoint. If we clear the session before, it
comes out empty and ProConnect errors.

Instead, add a dedicated post-logout action to clear the PC once
they're successfully logged out.