chore: rename Audit to Evaluation
A couple months ago someone raised a valid point: an audit is a
verified control of something, done and certified by an expert: our
service is none of that as it simply allows people to
check/self-assess whether their product fits into our standards.

Since semantices are important, rename "audit" to "evaluation".