evaluation: don't memoize the standards
This allows hot-reloading the standards without rebooting the server
in development. It also means re-reading a file (IO) in production
every time an evaluation starts but given the scope of this tool and
the probably of having 250+ services start an Evaluation at the same
time I don't see it being a problem.

We can always add an if-else around Rails.env but I try and avoid
these whenever I can.